revise the parameter 'Tag_FLOWIN_INDEX', 'Tag_FLOWOUT_INDEX' and 'Tag_ROUTING_LAYERS' to solve the problem: Failed reading file 0_flowin_index

+++ seims/src/seims_main/modules/hydrology/SUR_SGA/StormGreenAmpt.h
@@ -32,7 +32,7 @@ public:
 
     ~StormGreenAmpt();
 
-    void SetValue(const char* key, float value) OVERRIDE;
+    void SetValue(const char* key, int value) OVERRIDE;
 
     void Set1DData(const char* key, int n, float* data) OVERRIDE;
